@@ -3,29 +3,28 @@
 
 using System.Threading.Tasks;
 
-namespace System.Device.Gpio.Tests
+namespace System.Device.Gpio.Tests;
+
+public static class TimeoutHelper
 {
-    public static class TimeoutHelper
+    public static void CompletesInTime(Action test, TimeSpan timeout)
     {
-        public static void CompletesInTime(Action test, TimeSpan timeout)
+        Task task = Task.Run(test);
+        bool completedInTime = Task.WaitAll(new[] { task }, timeout);
+
+        if (task.Exception != null)
         {
-            Task task = Task.Run(test);
-            bool completedInTime = Task.WaitAll(new[] { task }, timeout);
-
-            if (task.Exception != null)
+            if (task.Exception.InnerExceptions.Count == 1)
             {
-                if (task.Exception.InnerExceptions.Count == 1)
-                {
-                    throw task.Exception.InnerExceptions[0];
-                }
-
-                throw task.Exception;
+                throw task.Exception.InnerExceptions[0];
             }
 
-            if (!completedInTime)
-            {
-                throw new TimeoutException($"Test did not complete in the specified timeout: {timeout.TotalSeconds} seconds.");
-            }
+            throw task.Exception;
+        }
+
+        if (!completedInTime)
+        {
+            throw new TimeoutException($"Test did not complete in the specified timeout: {timeout.TotalSeconds} seconds.");
         }
     }
 }
