@@ -5,18 +5,17 @@ using System.Device.Gpio.Drivers;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace System.Device.Gpio.Tests
+namespace System.Device.Gpio.Tests;
+
+[Trait("SkipOnTestRun", "Unix")]
+public class WindowsDriverTests : GpioControllerTestBase
 {
-    [Trait("SkipOnTestRun", "Unix")]
-    public class WindowsDriverTests : GpioControllerTestBase
+    public WindowsDriverTests(ITestOutputHelper testOutputHelper)
+        : base(testOutputHelper)
     {
-        public WindowsDriverTests(ITestOutputHelper testOutputHelper)
-            : base(testOutputHelper)
-        {
-        }
-
-        protected override GpioDriver GetTestDriver() => new Windows10Driver();
-
-        protected override PinNumberingScheme GetTestNumberingScheme() => PinNumberingScheme.Logical;
     }
+
+    protected override GpioDriver GetTestDriver() => new Windows10Driver();
+
+    protected override PinNumberingScheme GetTestNumberingScheme() => PinNumberingScheme.Logical;
 }
