@@ -3,24 +3,23 @@
 
 using System.Runtime.InteropServices;
 
-namespace System.Device.Gpio
+namespace System.Device.Gpio;
+
+/// <summary>
+/// Pointer to a general-purpose I/O (GPIO) chip.
+/// </summary>
+internal class SafeChipHandle : SafeHandle
 {
-    /// <summary>
-    /// Pointer to a general-purpose I/O (GPIO) chip.
-    /// </summary>
-    internal class SafeChipHandle : SafeHandle
+    public SafeChipHandle()
+        : base(IntPtr.Zero, true)
     {
-        public SafeChipHandle()
-            : base(IntPtr.Zero, true)
-        {
-        }
-
-        protected override bool ReleaseHandle()
-        {
-            Interop.libgpiod.gpiod_chip_close(handle);
-            return true;
-        }
-
-        public override bool IsInvalid => handle == IntPtr.Zero || handle == Interop.libgpiod.InvalidHandleValue;
     }
+
+    protected override bool ReleaseHandle()
+    {
+        Interop.libgpiod.gpiod_chip_close(handle);
+        return true;
+    }
+
+    public override bool IsInvalid => handle == IntPtr.Zero || handle == Interop.libgpiod.InvalidHandleValue;
 }
