@@ -4,108 +4,107 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace System.Device.Gpio.Drivers
+namespace System.Device.Gpio.Drivers;
+
+/// <summary>
+/// A GPIO driver for the HummingBoard.
+/// </summary>
+public class HummingBoardDriver : GpioDriver
 {
     /// <summary>
-    /// A GPIO driver for the HummingBoard.
+    /// We can't have HummingBoardDriver inherit from Windows10Driver since that would cause issues
+    /// when we decide to implement the driver on Linux. Having an internal driver instead that we
+    /// will forward the calls for now.
     /// </summary>
-    public class HummingBoardDriver : GpioDriver
+    private Windows10Driver _internalDriver;
+
+    /// <summary>
+    /// Constructs a new instance of a <see cref="HummingBoardDriver"/>.
+    /// This is currently only supported on Windows.
+    /// </summary>
+    /// <exception cref="PlatformNotSupportedException">The current operating system is not Windows</exception>
+    public HummingBoardDriver()
     {
-        /// <summary>
-        /// We can't have HummingBoardDriver inherit from Windows10Driver since that would cause issues
-        /// when we decide to implement the driver on Linux. Having an internal driver instead that we
-        /// will forward the calls for now.
-        /// </summary>
-        private Windows10Driver _internalDriver;
-
-        /// <summary>
-        /// Constructs a new instance of a <see cref="HummingBoardDriver"/>.
-        /// This is currently only supported on Windows.
-        /// </summary>
-        /// <exception cref="PlatformNotSupportedException">The current operating system is not Windows</exception>
-        public HummingBoardDriver()
+        if (Environment.OSVersion.Platform != PlatformID.Win32NT)
         {
-            if (Environment.OSVersion.Platform != PlatformID.Win32NT)
-            {
-                throw new PlatformNotSupportedException($"{GetType().Name} is only supported on Windows.");
-            }
-
-            _internalDriver = new Windows10Driver();
+            throw new PlatformNotSupportedException($"{GetType().Name} is only supported on Windows.");
         }
 
-        /// <inheritdoc/>
-        protected internal override int PinCount => _internalDriver.PinCount;
+        _internalDriver = new Windows10Driver();
+    }
 
-        /// <inheritdoc/>
-        protected internal override void AddCallbackForPinValueChangedEvent(int pinNumber, PinEventTypes eventTypes, PinChangeEventHandler callback) => _internalDriver.AddCallbackForPinValueChangedEvent(pinNumber, eventTypes, callback);
+    /// <inheritdoc/>
+    protected internal override int PinCount => _internalDriver.PinCount;
 
-        /// <inheritdoc/>
-        protected internal override void ClosePin(int pinNumber) => _internalDriver.ClosePin(pinNumber);
+    /// <inheritdoc/>
+    protected internal override void AddCallbackForPinValueChangedEvent(int pinNumber, PinEventTypes eventTypes, PinChangeEventHandler callback) => _internalDriver.AddCallbackForPinValueChangedEvent(pinNumber, eventTypes, callback);
 
-        /// <summary>
-        /// Converts a board pin number to the driver's logical numbering scheme.
-        /// </summary>
-        /// <param name="pinNumber">The board pin number to convert.</param>
-        /// <returns>The pin number in the driver's logical numbering scheme.</returns>
-        protected internal override int ConvertPinNumberToLogicalNumberingScheme(int pinNumber)
+    /// <inheritdoc/>
+    protected internal override void ClosePin(int pinNumber) => _internalDriver.ClosePin(pinNumber);
+
+    /// <summary>
+    /// Converts a board pin number to the driver's logical numbering scheme.
+    /// </summary>
+    /// <param name="pinNumber">The board pin number to convert.</param>
+    /// <returns>The pin number in the driver's logical numbering scheme.</returns>
+    protected internal override int ConvertPinNumberToLogicalNumberingScheme(int pinNumber)
+    {
+        return pinNumber switch
         {
-            return pinNumber switch
-            {
-                7 => 1,
-                11 => 73,
-                12 => 72,
-                13 => 71,
-                15 => 10,
-                16 => 194,
-                18 => 195,
-                22 => 67,
-                _ => throw new ArgumentException($"Board (header) pin {pinNumber} is not a GPIO pin on the {GetType().Name} device.", nameof(pinNumber))
-            };
-        }
+            7 => 1,
+            11 => 73,
+            12 => 72,
+            13 => 71,
+            15 => 10,
+            16 => 194,
+            18 => 195,
+            22 => 67,
+            _ => throw new ArgumentException($"Board (header) pin {pinNumber} is not a GPIO pin on the {GetType().Name} device.", nameof(pinNumber))
+        };
+    }
 
-        /// <summary>
-        /// Gets the mode of a pin.
-        /// </summary>
-        /// <param name="pinNumber">The pin number in the driver's logical numbering scheme.</param>
-        /// <returns>The mode of the pin.</returns>
-        protected internal override PinMode GetPinMode(int pinNumber)
-        {
-            throw new NotImplementedException();
-        }
+    /// <summary>
+    /// Gets the mode of a pin.
+    /// </summary>
+    /// <param name="pinNumber">The pin number in the driver's logical numbering scheme.</param>
+    /// <returns>The mode of the pin.</returns>
+    protected internal override PinMode GetPinMode(int pinNumber)
+    {
+        throw new NotImplementedException();
+    }
 
-        /// <inheritdoc/>
-        protected internal override bool IsPinModeSupported(int pinNumber, PinMode mode) => _internalDriver.IsPinModeSupported(pinNumber, mode);
+    /// <inheritdoc/>
+    protected internal override bool IsPinModeSupported(int pinNumber, PinMode mode) => _internalDriver.IsPinModeSupported(pinNumber, mode);
 
-        /// <inheritdoc/>
-        protected internal override void OpenPin(int pinNumber) => _internalDriver.OpenPin(pinNumber);
+    /// <inheritdoc/>
+    protected internal override void OpenPin(int pinNumber) => _internalDriver.OpenPin(pinNumber);
 
-        /// <inheritdoc/>
-        protected internal override PinValue Read(int pinNumber) => _internalDriver.Read(pinNumber);
+    /// <inheritdoc/>
+    protected internal override PinValue Read(int pinNumber) => _internalDriver.Read(pinNumber);
 
-        /// <inheritdoc/>
-        protected internal override void RemoveCallbackForPinValueChangedEvent(int pinNumber, PinChangeEventHandler callback) => _internalDriver.RemoveCallbackForPinValueChangedEvent(pinNumber, callback);
+    /// <inheritdoc/>
+    protected internal override void RemoveCallbackForPinValueChangedEvent(int pinNumber, PinChangeEventHandler callback) => _internalDriver.RemoveCallbackForPinValueChangedEvent(pinNumber, callback);
 
-        /// <inheritdoc/>
-        protected internal override void SetPinMode(int pinNumber, PinMode mode) => _internalDriver.SetPinMode(pinNumber, mode);
+    /// <inheritdoc/>
+    protected internal override void SetPinMode(int pinNumber, PinMode mode) => _internalDriver.SetPinMode(pinNumber, mode);
 
-        /// <inheritdoc/>
-        protected internal override void SetPinMode(int pinNumber, PinMode mode, PinValue initialValue) => _internalDriver.SetPinMode(pinNumber, mode, initialValue);
+    /// <inheritdoc/>
+    protected internal override void SetPinMode(int pinNumber, PinMode mode, PinValue initialValue) => _internalDriver.SetPinMode(pinNumber, mode, initialValue);
 
-        /// <inheritdoc/>
-        protected internal override WaitForEventResult WaitForEvent(int pinNumber, PinEventTypes eventTypes, CancellationToken cancellationToken) => _internalDriver.WaitForEvent(pinNumber, eventTypes, cancellationToken);
+    /// <inheritdoc/>
+    protected internal override WaitForEventResult WaitForEvent(int pinNumber, PinEventTypes eventTypes, CancellationToken cancellationToken) => _internalDriver.WaitForEvent(pinNumber, eventTypes, cancellationToken);
 
-        /// <inheritdoc/>
-        protected internal override ValueTask<WaitForEventResult> WaitForEventAsync(int pinNumber, PinEventTypes eventTypes, CancellationToken cancellationToken) => _internalDriver.WaitForEventAsync(pinNumber, eventTypes, cancellationToken);
+    /// <inheritdoc/>
+    protected internal override ValueTask<WaitForEventResult> WaitForEventAsync(int pinNumber, PinEventTypes eventTypes, CancellationToken cancellationToken) => _internalDriver.WaitForEventAsync(pinNumber, eventTypes, cancellationToken);
 
-        /// <inheritdoc/>
-        protected internal override void Write(int pinNumber, PinValue value) => _internalDriver.Write(pinNumber, value);
+    /// <inheritdoc/>
+    protected internal override void Write(int pinNumber, PinValue value) => _internalDriver.Write(pinNumber, value);
 
-        /// <inheritdoc/>
-        protected override void Dispose(bool disposing)
-        {
-            _internalDriver?.Dispose();
-            _internalDriver = null!;
-            base.Dispose(disposing);
-        }
+    /// <inheritdoc/>
+    protected override void Dispose(bool disposing)
+    {
+        _internalDriver?.Dispose();
+        _internalDriver = null!;
+        base.Dispose(disposing);
     }
 }
