@@ -1,21 +1,20 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace System.Device.Gpio
+namespace System.Device.Gpio;
+
+/// <summary>
+/// Different numbering schemes supported by GPIO controllers and drivers.
+/// </summary>
+public enum PinNumberingScheme
 {
     /// <summary>
-    /// Different numbering schemes supported by GPIO controllers and drivers.
+    /// The logical representation of the GPIOs. Refer to the microcontroller's datasheet to find this information.
     /// </summary>
-    public enum PinNumberingScheme
-    {
-        /// <summary>
-        /// The logical representation of the GPIOs. Refer to the microcontroller's datasheet to find this information.
-        /// </summary>
-        Logical,
+    Logical,
 
-        /// <summary>
-        /// The physical pin numbering that is usually accessible by the board headers.
-        /// </summary>
-        Board
-    }
+    /// <summary>
+    /// The physical pin numbering that is usually accessible by the board headers.
+    /// </summary>
+    Board
 }
