@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System;
+using System.Collections.Generic;
 using System.Device;
 using System.Device.I2c;
 using Iot.Device.Common;
@@ -23,41 +24,6 @@ namespace Iot.Device.DHTxx
         private const byte DHT10_CMD_START = 0b_1010_1100;
         private const byte DHT10_CMD_SOFTRESET = 0b_1011_1010;
 
-        // state, humi[20-13], humi[12-5], humi[4-1]temp[20-17], temp[16-9], temp[8-1]
-        private ValueArray<byte> _dht10ReadBuff = new ValueArray<byte>(6);
-
-        /// <summary>
-        /// Get the last read of relative humidity in percentage
-        /// </summary>
-        /// <remarks>
-        /// If last read was not successfull, it returns double.NaN
-        /// </remarks>
-        [Obsolete("This property will be removed in a future release. Use TryGetHumidity instead.")]
-        public override RelativeHumidity Humidity
-        {
-            get
-            {
-                ReadData();
-                return GetHumidity(_dht10ReadBuff.AsSpan());
-            }
-        }
-
-        /// <summary>
-        /// Get the last read temperature
-        /// </summary>
-        /// <remarks>
-        /// If last read was not successfull, it returns double.NaN
-        /// </remarks>
-        [Obsolete("This property will be removed in a future release. Use TryGetTemperature instead.")]
-        public override Temperature Temperature
-        {
-            get
-            {
-                ReadData();
-                return GetTemperature(_dht10ReadBuff.AsSpan());
-            }
-        }
-
         /// <summary>
         /// Create a DHT10 sensor through I2C
         /// </summary>
@@ -71,23 +37,24 @@ namespace Iot.Device.DHTxx
             i2cDevice.WriteByte(DHT10_CMD_INIT);
         }
 
-        internal override ValueArray<byte> ReadThroughI2c()
+        internal override byte[] ReadThroughI2c()
         {
             if (_i2cDevice is null)
             {
                 throw new Exception("I2C decvice not configured.");
             }
 
-            // DHT10 has no calibration bits
+            // DHT10 has no checksum bits
             _isLastReadSuccessful = true;
 
             _i2cDevice.WriteByte(DHT10_CMD_START);
             // make sure DHT10 ends measurement (in the datasheet P7)
             DelayHelper.DelayMilliseconds(75, true);
 
-            _i2cDevice.Read(_dht10ReadBuff.AsSpan());
+            byte[] data = new byte[6];
+            _i2cDevice.Read(data.AsSpan());
 
-            return _dht10ReadBuff;
+            return data;
         }
 
         internal override RelativeHumidity GetHumidity(Span<byte> readBuff)
