@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System;
+using System.Collections.Generic;
 using System.Device;
 using System.Device.Gpio;
 using System.Device.I2c;
@@ -20,12 +21,13 @@ namespace Iot.Device.DHTxx
     [Interface("Temperature and Humidity Sensor DHTxx")]
     public abstract class DhtBase : IDisposable
     {
+        private const int BufferSize = 5;
         private readonly CommunicationProtocol _protocol;
 
         /// <summary>
         /// Read buffer
         /// </summary>
-        private ValueArray<byte> _readBuff = new ValueArray<byte>(5);
+        private byte[] _readBuff = new byte[BufferSize];
 
         /// <summary>
         /// GPIO pin
@@ -79,7 +81,7 @@ namespace Iot.Device.DHTxx
             get
             {
                 var buf = ReadData();
-                return IsLastReadSuccessful ? GetTemperature(buf.AsSpan()) : default(Temperature);
+                return IsLastReadSuccessful ? GetTemperature(buf) : default(Temperature);
             }
         }
 
@@ -95,7 +97,7 @@ namespace Iot.Device.DHTxx
             get
             {
                 var buf = ReadData();
-                return IsLastReadSuccessful ? GetHumidity(buf.AsSpan()) : default(RelativeHumidity);
+                return IsLastReadSuccessful ? GetHumidity(buf) : default(RelativeHumidity);
             }
         }
 
@@ -152,9 +154,10 @@ namespace Iot.Device.DHTxx
         }
 
         /// <summary>
-        /// Start a reading
+        /// Read data from the sensor. Returns the last read value if the last read operation was too close.
         /// </summary>
-        internal virtual ValueArray<byte> ReadData()
+        /// <returns>The raw data last read</returns>
+        internal virtual byte[] ReadData()
         {
             // The time of two measurements should be more than 1s.
             if (Environment.TickCount - _lastMeasurement < MinTimeBetweenReads.Milliseconds)
@@ -164,18 +167,20 @@ namespace Iot.Device.DHTxx
 
             if (_protocol == CommunicationProtocol.OneWire)
             {
-                return ReadThroughOneWire();
+                _readBuff = ReadThroughOneWire();
             }
             else
             {
-                return ReadThroughI2c();
+                _readBuff = ReadThroughI2c();
             }
+
+            return _readBuff;
         }
 
         /// <summary>
         /// Read through One-Wire
         /// </summary>
-        internal virtual ValueArray<byte> ReadThroughOneWire()
+        internal virtual byte[] ReadThroughOneWire()
         {
             if (_controller is null)
             {
@@ -206,12 +211,13 @@ namespace Iot.Device.DHTxx
 
             // DHT corresponding signal - LOW - about 80 microseconds
             count = _loopCount;
+            byte[] newBuf = new byte[BufferSize];
             while (_controller.Read(_pin) == PinValue.Low)
             {
                 if (count-- == 0)
                 {
                     _isLastReadSuccessful = false;
-                    return default;
+                    return newBuf;
                 }
             }
 
@@ -222,7 +228,7 @@ namespace Iot.Device.DHTxx
                 if (count-- == 0)
                 {
                     _isLastReadSuccessful = false;
-                    return default;
+                    return newBuf;
                 }
             }
 
@@ -236,7 +242,7 @@ namespace Iot.Device.DHTxx
                     if (count-- == 0)
                     {
                         _isLastReadSuccessful = false;
-                        return default;
+                        return newBuf;
                     }
                 }
 
@@ -249,7 +255,7 @@ namespace Iot.Device.DHTxx
                     if (count-- == 0)
                     {
                         _isLastReadSuccessful = false;
-                        return default;
+                        return newBuf;
                     }
                 }
 
@@ -266,29 +272,28 @@ namespace Iot.Device.DHTxx
 
                 if (((i + 1) % 8) == 0)
                 {
-                    _readBuff[i / 8] = readVal;
+                    newBuf[i / 8] = readVal;
                 }
             }
 
             _lastMeasurement = Environment.TickCount;
 
-            if ((_readBuff[4] == ((_readBuff[0] + _readBuff[1] + _readBuff[2] + _readBuff[3]) & 0xFF)))
+            if ((newBuf[4] == ((newBuf[0] + newBuf[1] + newBuf[2] + newBuf[3]) & 0xFF)))
             {
-                _isLastReadSuccessful = (_readBuff[0] != 0) || (_readBuff[2] != 0);
+                _isLastReadSuccessful = (newBuf[0] != 0) || (newBuf[2] != 0);
             }
             else
             {
                 _isLastReadSuccessful = false;
-                return default;
             }
 
-            return _readBuff;
+            return newBuf;
         }
 
         /// <summary>
         /// Read through I2C
         /// </summary>
-        internal virtual ValueArray<byte> ReadThroughI2c()
+        internal virtual byte[] ReadThroughI2c()
         {
             if (_i2cDevice is null)
             {
@@ -298,21 +303,21 @@ namespace Iot.Device.DHTxx
             // DHT12 Humidity Register
             _i2cDevice.WriteByte(0x00);
             // humidity int, humidity decimal, temperature int, temperature decimal, checksum
-            _i2cDevice.Read(_readBuff.AsSpan());
+            byte[] newBuf = new byte[BufferSize];
+            _i2cDevice.Read(newBuf.AsSpan());
 
             _lastMeasurement = Environment.TickCount;
 
-            if ((_readBuff[4] == ((_readBuff[0] + _readBuff[1] + _readBuff[2] + _readBuff[3]) & 0xFF)))
+            if ((newBuf[4] == ((newBuf[0] + newBuf[1] + newBuf[2] + newBuf[3]) & 0xFF)))
             {
-                _isLastReadSuccessful = (_readBuff[0] != 0) || (_readBuff[2] != 0);
+                _isLastReadSuccessful = (newBuf[0] != 0) || (newBuf[2] != 0);
             }
             else
             {
                 _isLastReadSuccessful = false;
-                return default;
             }
 
-            return _readBuff;
+            return newBuf;
         }
 
         /// <summary>
@@ -331,7 +336,7 @@ namespace Iot.Device.DHTxx
             var buf = ReadData();
             if (_isLastReadSuccessful)
             {
-                temperature = GetTemperature(buf.AsSpan());
+                temperature = GetTemperature(buf);
                 return true;
             }
 
@@ -354,7 +359,7 @@ namespace Iot.Device.DHTxx
             var buf = ReadData();
             if (_isLastReadSuccessful)
             {
-                humidity = GetHumidity(buf.AsSpan());
+                humidity = GetHumidity(buf);
                 return true;
             }
 
